refactor: 파일 책임 분리 및 dotenv -> settings 로 변경
