+++ config.py
@@ -1,0 +1,19 @@
+from pathlib import Path
+
+from pydantic_settings import BaseSettings, SettingsConfigDict
+
+
+class Settings(BaseSettings):
+    model_config = SettingsConfigDict(
+        env_file=".env",
+        env_file_encoding="utf-8",
+    )
+
+    secret_key: str
+    password_pepper: str
+    algorithm: str = "HS256"
+    access_token_expire_minutes: int = 30
+    users_file: Path = Path("data/users.json")
+
+
+settings = Settings()
