+++ main.py
@@ -1,9 +1,5 @@
 from fastapi import FastAPI
 
-from dotenv import load_dotenv
-
-load_dotenv()
-
 from routers import users, posts, comments, likes
 
 app = FastAPI()
