+++ pyproject.toml
@@ -8,7 +8,7 @@ dependencies = [
     "pydantic[email]>=2.0.0,<3.0.0",
     "bcrypt>=4.0.0,<5.0.0",
     "pyjwt>=2.10.1",
-    "python-dotenv>=1.2.1,<2.0.0",
+    "pydantic-settings>=2.12.0",
 ]
 
 [project.optional-dependencies]
